+++ packages/mds-db/index.ts
@@ -1364,7 +1364,7 @@ async function readTripIds(params: Partial<ReadTripIdsQueryParams> = {}): Promis
   }
 }
 
-export default {
+export = {
   initialize,
   health,
   seed,
