+++ packages/mds-provider/event-processor/index.ts
@@ -17,10 +17,11 @@
 import db from '@mds-core/mds-db'
 import { promisify } from 'util'
 import logger from '@mds-core/mds-logger'
-import stream, { ReadStreamOptions, StreamItem } from '@mds-core/mds-stream'
+import stream from '@mds-core/mds-stream'
 import uuid from 'uuid'
 import { isUUID } from '@mds-core/mds-utils'
 import { VehicleEvent } from '@mds-core/mds-types'
+import { ReadStreamOptions, StreamItem } from '@mds-core/mds-stream/types'
 import { DeviceLabeler } from './labelers/device-labeler'
 import { ProviderLabeler } from './labelers/provider-labeler'
 import { StreamEntry } from './types'
