+++ packages/mds-stream/types.ts
@@ -1,0 +1,15 @@
+export type ReadStreamOptions = Partial<{
+  count: number
+  block: number
+  noack: boolean
+}>
+
+export const Streams = ['device:index', 'device:raw', 'provider:event'] as const
+export const [DEVICE_INDEX_STREAM, DEVICE_RAW_STREAM, PROVIDER_EVENT_STREAM] = Streams
+export type Stream = typeof Streams[number]
+
+export type StreamItemID = string
+export type StreamItemType = string
+export type StreamItemData = string
+export type StreamItem = [StreamItemID, [StreamItemType, StreamItemData]]
+export type ReadStreamResult = [Stream, StreamItem[]]
